move bin/ scripts to partis/scripts/ for editable pip install
With scripts=[] in setup.py, pip copies bin/ scripts into the venv, so
edits don't take effect without reinstalling. Replace with console_scripts
entry points that import from a new partis/scripts/ package, where the
actual code now lives. bin/ scripts become thin wrappers, and test/test.py
likewise. Also update partis/main.py to use runpy instead of subprocess.
